move getBlobAndCheckETags into executeTask
